refactor: move get_cache_values method
