+++ BSoup/hh_search.py
@@ -2,7 +2,8 @@ import os
 from pprint import pprint
 
 from bs4 import BeautifulSoup as Soup
-from helpers import get_response
+
+from helpers import get_response, get_cash_values
 
 
 URL = 'https://hh.ru/search/vacancy'
@@ -26,48 +27,6 @@ request_data = {
     'params': params,
     'headers': headers,
 }
-
-
-def get_cash_values(raw_data) -> tuple:
-    """
-    Parse input string to tuple with 3 elements.
-    If some of this elements is not exist, define as None
-
-    Args -> str:
-        input_string: string like '100 - 10000 RUB' or something like that
-
-    Returns:
-        tuple of min salary, max salary and kind of currency; None for each if empty
-    """
-    raw_data = raw_data.replace('<!-- -->', '')
-    max_salary = None
-    min_salary = None
-    currency = None
-
-    sep_position = raw_data.find('–')
-    if sep_position != -1:
-        min_salary = int(''.join([raw_data[i] for i in range(sep_position, -1, -1) if raw_data[i].isdigit()][::-1]))
-        max_salary = int(''.join([raw_data[i] for i in range(sep_position, len(raw_data)) if raw_data[i].isdigit()]))
-        sep_position = raw_data.rfind(' ')
-        currency = ''.join([raw_data[i] for i in range(sep_position, len(raw_data)) if raw_data[i].isalpha()])
-        currency = currency if len(currency) > 0 else None
-    else:
-        sep_position = raw_data.rfind(' ')
-        if sep_position != -1:
-            if raw_data.find('от') != -1:
-                min_salary = int(''.join([raw_data[i] for i in range(sep_position, -1, -1) if raw_data[i].isdigit()][::-1]))
-            if raw_data.find('до') != -1:
-                max_salary = int(''.join([raw_data[i] for i in range(sep_position, -1, -1) if raw_data[i].isdigit()][::-1]))
-            currency = ''.join([raw_data[i] for i in range(sep_position, len(raw_data)) if raw_data[i].isalpha()])
-            currency = currency if len(currency) > 0 else None
-        elif len(raw_data) > 0:
-            max_salary = ''.join([raw_data[i] for i in range(0, -1, -1) if raw_data[i].isdigit()][::-1])
-            max_salary = int(max_salary) if len(max_salary) > 0 else None
-            min_salary = max_salary
-        else:
-            min_salary = max_salary = currency = None
-    return min_salary, max_salary, currency
-
 
 if not os.path.exists('./pages/'):
     os.mkdir('./pages/')
